pyproject.toml
[tool.poetry]
name = "etl-snowflake-lineage-purview"
version = "0.1.0"
description = "Synchronisation du lineage Snowflake interne avec Purview via l'API Apache Atlas Publique"
authors = ["Mathis Fasolo <mathis.fasolo@soyhuce.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
pyapacheatlas = "^0.16.0"
snowflake-connector-python = {extras = ["pandas"], version = "^3.12.0"}
pandas = "^2.2.2"
pydantic-settings = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
